+++ lang/fr.go
@@ -5,8 +5,8 @@ var FR_STRINGS = LangString{
 	"output-desc":      "Répertoire de sortie",
 	"output-dir-empty": "Le répertoire de sortie est vide",
 	"silent":           "Désactiver l'impression du journal sur stdout",
-	"hello-world":      "Bonjour, monde %s\n",
 	"no-args":          "Aucun argument fourni",
 	"no-command":       "Aucune commande fournie",
 	"text-usage":       "Générer du texte à partir d'un prompt",
+	"unknown-sdk":      "Sdk inconnu \"%s\"",
 }