+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 ### Added
 
+* Add sdk for openai text completion
 * Add localized for `en` and `fr`
 * Read config.ini file